requirements.txt
mlflow
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv